BREAKING CHANGE: do not load/use unicode symbol table (#50)
Mapping ✊ to `raised-fist` by default is cute, but not particularly
useful and loading that enormous unicode table is not what people want
by default, I have to imagine. We can document how to add it in for
people who want it, but the primary use case requires a small library
that does languages. Omitting symbols is a good default.