feat: make Base64 fallback in browser far more robust
The Base64 fallback would throw on lone surrogates in the string. The
Node.js Base64 implementation strips out lone surrogates. Let's emulate
that on the browser so that we can hopefully return a usable (but not
SEO-friendly) slug when we don't have any of the characters in a string
in our charmap.